pyproject.toml
[project]
name = "py_ghst_tools"
version = "0.0.1"
description = "A collection of cadaverous World of Warcraft datamining tools"
authors = [
    { name="Ghostopheles", email="ghost@ghst.tools"}
]
license = {file = "LICENSE"}
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",

    "Intended Audience :: Developers",

    "License :: Public Domain",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">= 3.11"
dependencies = ["httpx[http2]", "typer", "pillow", "customtkinter"]

[project.scripts]
ghst = "ghst.tools.cli:handle_cli"

[project.urls]
Homepage = "https://github.com/Ghostopheles/py-ghst-tools"
Repository = "https://github.com/Ghostopheles/py-ghst-tools.git"
Issues = "https://github.com/Ghostopheles/py-ghst-tools/issues"

[build-system]
requires = ["setuptools", "wheel", "cibuildwheel"]
build-backend = "setuptools.build_meta"